LibWeb/CSS: Populate 'cached_x_samples' properly in 'CubicBezier'
When we populate 'm_cached_x_samples' the last value might
be less than 1.0, which leads to a crash when 'input_progress == 1'
since the binary search populate 'nearby_index' with the index of the
last element in 'm_cached_x_samples' and when we try to linearly
interpolate between 'm_cached_x_samples[nearby_index]' and
m_cached_x_samples[nearby_index + 2] we crash because of an index out
of bounds assertion failure.

This change makes sure the loop responsible for populating
'm_cached_x_samples' reaches 'input_progress'.